requirements.txt
python-telegram-bot==20.7
pyautogui==0.9.54
pillow==10.2.0
schedule==1.2.1 